Move Pathfinder to main.algorithms package
